Move Notion coordinator to its own module
